+++ endstone/event/__init__.py
@@ -87,6 +87,8 @@ __getattr__, __dir__, __all__ = lazy.attach(
             "BlockPistonRetractEvent",
             "BlockPlaceEvent",
             "BroadcastMessageEvent",
+            "StructureGrowEvent",
+            "TreeType",
             "ChunkEvent",
             "ChunkLoadEvent",
             "ChunkUnloadEvent",